refactor(cli): split goal controller helpers
